Move some session handling outside specific integration
