+++ src/main/integrations/main-process-session.ts
@@ -1,22 +1,8 @@
-import { getCurrentHub } from '@sentry/core';
-import { flush, NodeClient } from '@sentry/node';
-import { Integration, SessionContext, SessionStatus } from '@sentry/types';
+import { Integration } from '@sentry/types';
 import { logger } from '@sentry/utils';
 import { app } from 'electron';
-import { join } from 'path';
 
-import { ElectronNetTransport } from '../transports/electron-net';
-import { Store } from './store';
-
-const TERMINAL_STATES = [SessionStatus.Exited, SessionStatus.Crashed];
-
-const sessionStore: Store<SessionContext | undefined> = new Store(
-  join(app.getPath('userData'), 'sentry'),
-  'session',
-  undefined,
-);
-
-let previousSession = sessionStore.get();
+import { endSession, startSession } from '../sessions';
 
 /** Tracks sessions as the main process lifetime. */
 export class MainProcessSession implements Integration {
@@ -28,11 +14,7 @@ export class MainProcessSession implements Integration {
 
   /** @inheritDoc */
   public setupOnce(): void {
-    const hub = getCurrentHub();
-
-    logger.log('[MainProcessSession] Start session');
-    const newSession = hub.startSession();
-    sessionStore.set(newSession);
+    startSession();
 
     // We track sessions via the 'will-quit' event which is the last event emitted before close.
     //
@@ -61,63 +43,11 @@ export class MainProcessSession implements Integration {
 
     // Stop the exit so we have time to send the session
     event.preventDefault();
-    const hub = getCurrentHub();
 
-    const session = hub.getScope()?.getSession();
-
-    if (session && !TERMINAL_STATES.includes(session.status)) {
-      logger.log('[MainProcessSession] Ending session');
-      hub.endSession();
-    } else {
-      logger.log('[MainProcessSession] Session was already ended');
-    }
-
-    sessionStore.set(undefined, true);
-
-    await flush();
+    // End the session
+    await endSession();
 
     // After flush we can safely exit
     app.exit();
   };
-}
-
-/** Checks if the previous session needs sending as crashed or abnormal  */
-export async function checkPreviousSession(crashed: boolean): Promise<void> {
-  if (previousSession) {
-    const status = crashed ? SessionStatus.Crashed : SessionStatus.Abnormal;
-
-    logger.log(`[MainProcessSession] Previous session ${status}}`);
-    // eslint-disable-next-line @typescript-eslint/no-unsafe-member-access
-    const transport = (getCurrentHub().getClient<NodeClient>() as any)
-      ._getBackend()
-      .getTransport() as ElectronNetTransport;
-
-    previousSession.status = status;
-    previousSession.errors = (previousSession.errors || 0) + 1;
-
-    await transport.sendSession(previousSession);
-
-    previousSession = undefined;
-  }
-}
-
-/** Sets the current session as crashed */
-export function sessionCrashed(): void {
-  logger.log('[MainProcessSession] Session Crashed');
-  const hub = getCurrentHub();
-  const session = hub.getScope()?.getSession();
-
-  if (!session) {
-    logger.log('[MainProcessSession] No session to update');
-    return;
-  }
-
-  if (!TERMINAL_STATES.includes(session.status)) {
-    logger.log(`[MainProcessSession] Setting session as crashed`);
-    session.update({ status: SessionStatus.Crashed, errors: (session.errors += 1) });
-  } else {
-    logger.log('[MainProcessSession] Session already ended');
-  }
-
-  hub.captureSession();
 }
