+++ src/main/store.ts
@@ -1,8 +1,8 @@
 import { logger } from '@sentry/utils';
-import { existsSync, readFileSync, rmSync, writeFileSync } from 'fs';
+import { existsSync, readFileSync, unlinkSync, writeFileSync } from 'fs';
 import { dirname, join } from 'path';
 
-import { mkdirpSync } from '../fs';
+import { mkdirpSync } from './fs';
 
 /**
  * Note, this class is only compatible with Node.
@@ -87,7 +87,7 @@ export class Store<T> {
   private _flush(): void {
     try {
       if (this._data === undefined) {
-        rmSync(this._path);
+        unlinkSync(this._path);
       } else {
         mkdirpSync(dirname(this._path));
         writeFileSync(this._path, JSON.stringify(this._data));
